+++ src/utils/common.js
@@ -1,5 +1,50 @@
 import moment from "moment";
 
+export const parseFormData = (formData, event, allOffers) => {
+  const selectedType = document.querySelector(`.event__type-output`)
+    .textContent.trim().split(` `)[0];
+
+  const checkedOffersTitle = Array.from(document.querySelectorAll(`.event__offer-checkbox`))
+      .filter((it) => it.checked)
+      .map((offer) => offer.getAttribute(`value`));
+
+  const getOffers = () => {
+    return allOffers.map((offer) => {
+      if (checkedOffersTitle.includes(offer.title)) {
+        return Object.assign({}, offer, {checked: true});
+      } else {
+        return Object.assign({}, offer, {checked: false});
+      }
+    });
+  };
+
+  const convertData = (dateString) => {
+    const monthIndex = dateString[3] + dateString[4];
+    const dayIndex = dateString[0] + dateString[1];
+    return monthIndex + `/` + dayIndex + dateString.slice(5);
+  };
+
+  return {
+    id: event.id,
+    eventType: {
+      name: selectedType,
+      offers: getOffers(),
+      group: event.eventType.group,
+
+    },
+    destination: {
+      name: formData.get(`event-destination`),
+      pictures: event.destination.pictures,
+      description: event.destination.description,
+    },
+
+    dateFrom: convertData(formData.get(`event-start-time`)),
+    dateTo: convertData(formData.get(`event-end-time`)),
+    basePrice: formData.get(`event-price`),
+    isFavorite: !!(formData.get(`event-favorite`)),
+  };
+};
+
 export const formatDate = (date) => {
   return moment(date).format(`L`);
 };
